Move _get_members_to_document
